chore: rename profile, 'GetAssetState' use case, MediaContainer model
