+++ grid/social-media/upload-from-url/maps/mock.suma
@@ -1,4 +1,4 @@
-profile = "social-media/upload-url@1.0"
+profile = "social-media/upload-from-url@1.0"
 provider = "mock"
 
 map RegisterUpload {
@@ -7,7 +7,7 @@ map RegisterUpload {
   }
 }
 
-map GetAssetState {
+map GetUploadState {
   return map result {
       state = 'Finished'
   }
