+++ superface/super.json
@@ -445,14 +445,14 @@
         }
       }
     },
-    "social-media/upload-url": {
-      "file": "../grid/social-media/upload-url/profile.supr",
+    "social-media/upload-from-url": {
+      "file": "../grid/social-media/upload-from-url/profile.supr",
       "providers": {
         "mock": {
-          "file": "../grid/social-media/upload-url/maps/mock.suma"
+          "file": "../grid/social-media/upload-from-url/maps/mock.suma"
         },
         "instagram": {
-          "file": "../grid/social-media/upload-url/maps/instagram.suma"
+          "file": "../grid/social-media/upload-from-url/maps/instagram.suma"
         }
       }
     },
